Some mockups,...

+++ recreate_database.py
@@ -1,7 +1,12 @@
 #!/usr/bin/env python
-
+from os import getenv
 from feeder.models import *
 
-Base.metadata.drop_all(db_connect())
+drop = getenv("DROP", "False")
+
+if drop is "True":
+    Base.metadata.drop_all(db_connect())
 
 Base.metadata.create_all(db_connect())
+
+print("Done.")
